Student Management System using Spring boot-Maven Project. Using Perform Curd Operation .Here Thyemeleaf MVC as a view Resolver.
Create database as "sms" in MySql database. and table name as "student".
The main aim is to design a’ student management system’. To be aware STUDENT list to the teacher. Using spring boot to create project output. Eclipse workspace without installing to local maven repository. Hibernate provides a framework for mapping an object-oriented domain model to the relational database. It generates SQL calls and relieves the developer from the result set's manual handling and object conversion.

this project is created based on
 JDK
Eclipse
JPA
Spring Boot
Mysql workbench