student management system
The main aim is to design a’ student management system’. To be aware STUDENT list to the teacher. Using spring boot to create project output. Eclipse workspace without installing to local maven repository. Hibernate provides a framework for mapping an object-oriented domain model to a relational database. It generates SQL calls and relieves the developer from the result set's manual handling and object conversion.